move WASM files to separate folder
