+++ src/xrpld/app/wasm/WasmHostFuncImpl.h
@@ -19,8 +19,8 @@
 
 #pragma once
 
-#include <xrpld/app/misc/WasmHostFunc.h>
 #include <xrpld/app/tx/detail/ApplyContext.h>
+#include <xrpld/app/wasm/WasmHostFunc.h>
 
 namespace ripple {
 class WasmHostFunctionsImpl : public HostFunctions
