+++ src/xrpld/app/wasm/WasmHostFuncWrapper.h
@@ -19,7 +19,7 @@
 
 #pragma once
 
-#include <xrpld/app/misc/WamrVM.h>
+#include <xrpld/app/wasm/WamrVM.h>
 
 namespace ripple {
